refactor: dissolve utils.rs and constants.rs, relocate the URL-validation tests
utils.rs held three unrelated helpers (cookie, URL, filesystem) and
constants.rs held three constants each with a single consumer; move every
piece next to the code that uses it. Pure restructuring — no behaviour
change.

- cookie.rs: takes with_cookie (adding the Cookie header to a request
  builder) from utils; the module now covers both building the header value
  and applying it to a request.
- archive_metadata.rs: takes the archive.org URL contract from utils
  (validate_archive_url and the compiled URL_REGEX) and the sole-consumer
  constants from constants.rs (URL_PATTERN, XML_DEBUG_TRUNCATE_LEN); the
  check_valid_pattern / check_invalid_pattern tests move in from the main.rs
  binary, next to the code they test, and the doctest path follows.
- fs.rs (new): the ensure_not_symlink write-through-a-pre-planted-symlink
  guard — shared by the .part/final-file paths and the _files.xml save —
  gets its own module instead of riding in utils.
- main.rs: USER_AGENT (its only consumer, build_client) becomes a
  module-private const; the two URL tests are gone (relocated).
- constants.rs, utils.rs: deleted; lib.rs drops both and registers fs.
- AGENTS.md: the architecture listing follows — fs.rs added,
  utils.rs/constants.rs removed, and the widened main/archive_metadata/cookie
  descriptions.

No behaviour change: the public API and on-screen output are untouched, and
the test total stays at 105 passing — check_valid_pattern and
check_invalid_pattern just relocate from the main.rs binary into
archive_metadata (fmt and clippy -D warnings clean).
