+++ src/lib.rs
@@ -6,16 +6,15 @@
 //! with support for resumable downloads and MD5 hash verification.
 
 pub mod archive_metadata;
-pub mod constants;
 pub mod cookie;
 pub mod display;
 pub mod downloader;
 pub mod error;
 pub mod filename;
+pub mod fs;
 pub mod plan;
 #[cfg(test)]
 pub mod test_support;
-pub mod utils;
 
 // Re-export the error types for convenience
 pub use error::{IaGetError, Result};
